+++ lib/fumanchu/parser.ex
@@ -1,23 +1,3 @@
-defmodule FuManchu.Parser.TokenMissingError do
-  defexception [:message]
-
-  # TODO: Support passing in the filename of the template
-  def exception(%{parsed_line: parsed_line, token_name: token_name, token: token, starting: starting, starting_line: starting_line}) do
-    message = ~s[template:#{parsed_line}: missing #{token_name}: #{inspect token} (for #{inspect starting} starting at line #{starting_line})]
-    %FuManchu.Parser.TokenMissingError{message: message}
-  end
-end
-
-defmodule FuManchu.Parser.TokenUnrecognizedError do
-  defexception [:message]
-
-  # TODO: Support passing in the filename of the template
-  def exception(%{line: line, token: token}) do
-    message = ~s[template:#{line}: unrecognized token: #{inspect token}]
-    %FuManchu.Parser.TokenUnrecognizedError{message: message}
-  end
-end
-
 defmodule FuManchu.Parser do
   alias FuManchu.Parser.TokenMissingError
   alias FuManchu.Parser.TokenUnrecognizedError
